+++ spec/lib/challenge_spec.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require "cli_helper"
 require "tmpdir"
 require "fileutils"
