+++ spec/lib/keylog_spec.rb
@@ -1,15 +1,15 @@
-require "spec_helper"
+require "cli_helper"
 
 include VimGolf
 
 describe VimGolf::Keylog do
 
-  Dir['spec/fixtures/*'].each do |f|
-    it "should parse #{f} logfile" do
+  Dir[File.join(File.dirname(__FILE__), 'fixtures', '*')].each do |f|
+    it "should parse #{File.basename(f)} logfile" do
       expect { Keylog.new(IO.read(f)).convert }.not_to raise_error
     end
 
-    it "should score #{f} logfile" do
+    it "should score #{File.basename(f)} logfile" do
       expect { Keylog.new(IO.read(f)).convert }.not_to raise_error
     end
   end
