+++ spec/spec_helper.rb
@@ -10,19 +10,6 @@ require 'rspec/rails'
 # in spec/support/ and its subdirectories.
 Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
 
-module Kernel
-  def capture_stdio(input = nil, &block)
-    org_stdin, $stdin = $stdin, StringIO.new(input) if input
-    org_stdout, $stdout = $stdout, StringIO.new
-    yield
-    return @out = $stdout.string
-  ensure
-    $stdout = org_stdout
-    $stdin = org_stdin
-  end
-  alias capture_stdout capture_stdio
-end
-
 RSpec.configure do |config|
   config.mock_with :rspec
   config.infer_spec_type_from_file_location!
